refactor: mdc 이넘 추가
