+++ backend/src/main/java/com/now/naaga/common/LogFilter.java
@@ -7,6 +7,9 @@ import org.slf4j.MDC;
 
 import java.io.IOException;
 
+import static com.now.naaga.common.MdcToken.QUERYCOUNT;
+import static com.now.naaga.common.MdcToken.TIME;
+
 public class LogFilter implements Filter {
 
     private final Logger log = LoggerFactory.getLogger(this.getClass().getSimpleName());
@@ -35,8 +38,8 @@ public class LogFilter implements Filter {
         final int queryCount = queryCounter.count();
         final long end = System.currentTimeMillis();
         final long time = end - start;
-        MDC.put("time", String.valueOf(time) + "ms");
-        MDC.put("queryCount", String.valueOf(queryCount));
+        MDC.put(TIME.getKey(), String.valueOf(time) + "ms");
+        MDC.put(QUERYCOUNT.getKey(), String.valueOf(queryCount));
         log.info("queryCount: {}", queryCount);
     }
