+++ backend/src/main/java/com/now/naaga/common/MdcToken.java
@@ -2,7 +2,10 @@ package com.now.naaga.common;
 
 public enum MdcToken {
     REQUEST_ID("request_id"),
-    ;
+    URI("uri"),
+    TIME("time"),
+    QUERYCOUNT("queryCount"),
+    METHOD("method");
 
     private final String key;
