+++ backend/src/main/java/com/now/naaga/common/config/FilterConfig.java
@@ -1,5 +1,8 @@
-package com.now.naaga.common;
+package com.now.naaga.common.config;
 
+import com.now.naaga.common.LogFilter;
+import com.now.naaga.common.MdcFilter;
+import com.now.naaga.common.QueryCounter;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
